introduce VehicleStatusUtils
