+++ bundles/org.openhab.binding.bmwconnecteddrive/src/main/java/org/openhab/binding/bmwconnecteddrive/internal/dto/status/VehicleStatus.java
@@ -12,15 +12,10 @@
  */
 package org.openhab.binding.bmwconnecteddrive.internal.dto.status;
 
-import static org.openhab.binding.bmwconnecteddrive.internal.utils.Constants.*;
-
-import java.lang.reflect.Field;
-import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
 
 import org.openhab.binding.bmwconnecteddrive.internal.utils.Constants;
-import org.openhab.binding.bmwconnecteddrive.internal.utils.Converter;
 
 import com.google.gson.annotations.SerializedName;
 
@@ -74,175 +69,4 @@ public class VehicleStatus {
     public boolean dcsCchOngoing;// ":false
     public List<CCMMessage> checkControlMessages = new ArrayList<CCMMessage>();// ":[],
     public List<CBSMessage> cbsData = new ArrayList<CBSMessage>();
-
-    /**
-     * Get Next Service for Date and / or Mileage
-     *
-     * @param imperial
-     * @return
-     */
-    public CBSMessage getNextService() {
-        CBSMessage cbs = new CBSMessage();
-        if (cbsData == null) {
-            return cbs;
-        }
-        if (cbsData.isEmpty()) {
-            return cbs;
-        } else {
-            int serviceMileage = Integer.MAX_VALUE;
-            LocalDateTime serviceDate = LocalDateTime.now().plusYears(100);
-
-            for (int i = 0; i < cbsData.size(); i++) {
-                CBSMessage entry = cbsData.get(i);
-                if (entry.cbsRemainingMileage != 0 && entry.cbsDueDate != null) {
-                    LocalDateTime d = LocalDateTime.parse(entry.cbsDueDate + Constants.UTC_APPENDIX);
-                    // LocalDate d = LocalDate.parse(entry.cbsDueDate + APPENDIX_DAY,
-                    // Converter.SERVICE_DATE_INPUT_PATTERN);
-                    if ((entry.cbsRemainingMileage < serviceMileage) || (d.isBefore(serviceDate))) {
-                        serviceDate = d;
-                        serviceMileage = entry.cbsRemainingMileage;
-                        cbs = entry;
-                    }
-                } else if (entry.cbsRemainingMileage != 0) {
-                    if (entry.cbsRemainingMileage < serviceMileage) {
-                        serviceMileage = entry.cbsRemainingMileage;
-                        cbs = entry;
-                    }
-                } else if (entry.cbsDueDate != null) {
-                    LocalDateTime d = LocalDateTime.parse(entry.cbsDueDate + Constants.UTC_APPENDIX);
-                    // LocalDate d = LocalDate.parse(entry.cbsDueDate + APPENDIX_DAY,
-                    // Converter.SERVICE_DATE_INPUT_PATTERN);
-                    if (d.isBefore(serviceDate)) {
-                        serviceDate = d;
-                        cbs = entry;
-                    }
-                }
-            }
-        }
-        return cbs;
-    }
-
-    public String getNextServiceDate() {
-        if (cbsData == null) {
-            return Constants.NULL_DATE;
-        }
-        if (cbsData.isEmpty()) {
-            return Constants.NULL_DATE;
-        } else {
-            LocalDateTime farFuture = LocalDateTime.now().plusYears(100);
-            LocalDateTime serviceDate = farFuture;
-            for (int i = 0; i < cbsData.size(); i++) {
-                CBSMessage entry = cbsData.get(i);
-                if (entry.cbsDueDate != null) {
-                    LocalDateTime d = LocalDateTime.parse(entry.cbsDueDate + Constants.UTC_APPENDIX);
-                    // LocalDate d = LocalDate.parse(entry.cbsDueDate + APPENDIX_DAY,
-                    // Converter.SERVICE_DATE_INPUT_PATTERN);
-                    if (d.isBefore(serviceDate)) {
-                        serviceDate = d;
-                    }
-                }
-            }
-            if (serviceDate.equals(farFuture)) {
-                return Constants.NULL_DATE;
-            } else {
-                return serviceDate.format(Converter.DATE_INPUT_PATTERN);
-            }
-        }
-    }
-
-    public int getNextServiceMileage() {
-        if (cbsData == null) {
-            return -1;
-        }
-        if (cbsData.isEmpty()) {
-            return -1;
-        } else {
-            int serviceMileage = Integer.MAX_VALUE;
-            for (int i = 0; i < cbsData.size(); i++) {
-                CBSMessage entry = cbsData.get(i);
-                if (entry.cbsRemainingMileage != -1) {
-                    if (entry.cbsRemainingMileage < serviceMileage) {
-                        serviceMileage = entry.cbsRemainingMileage;
-                    }
-                }
-            }
-            if (serviceMileage != Integer.MAX_VALUE) {
-                return serviceMileage;
-            } else {
-                return -1;
-            }
-        }
-    }
-
-    public String checkControlActive() {
-        if (checkControlMessages == null) {
-            return UNKNOWN;
-        }
-        if (checkControlMessages.isEmpty()) {
-            return NOT_ACTIVE;
-        } else {
-            return ACTIVE;
-        }
-    }
-
-    public String getCheckControl() {
-        if (checkControlMessages == null) {
-            return Converter.toTitleCase(UNKNOWN);
-        }
-        if (checkControlMessages.isEmpty()) {
-            return OK;
-        } else {
-            return Converter.toTitleCase(checkControlMessages.get(0).ccmDescriptionShort);
-        }
-    }
-
-    public String getUpdateTime() {
-        if (internalDataTimeUTC != null) {
-            return internalDataTimeUTC;
-        } else if (updateTime != null) {
-            return updateTime;
-        } else {
-            return Constants.NULL_DATE;
-        }
-    }
-
-    /**
-     * Check for certain Windows or Doors DTO object the "Closed" Status
-     * INVALID values will be ignored
-     *
-     * @param dto
-     * @return Closed if all "Closed", "Open" otherwise
-     */
-    public static String checkClosed(Object dto) {
-        String overallState = Constants.UNKNOWN;
-        for (Field field : dto.getClass().getDeclaredFields()) {
-            try {
-                Object d = field.get(dto);
-                if (d != null) {
-                    String state = d.toString();
-                    // skip invalid entries - they don't apply to this Vehicle
-                    if (!state.equalsIgnoreCase(INVALID)) {
-                        if (state.equalsIgnoreCase(OPEN)) {
-                            overallState = OPEN;
-                            // stop searching for more open items - overall Doors / Windows are OPEN
-                            break;
-                        } else if (state.equalsIgnoreCase(INTERMEDIATE)) {
-                            if (!overallState.equalsIgnoreCase(OPEN)) {
-                                overallState = INTERMEDIATE;
-                                // continue searching - maybe another Door / Window is OPEN
-                            }
-                        } else if (state.equalsIgnoreCase(CLOSED)) {
-                            // at least one valid object needs to be found in order to reply "CLOSED"
-                            if (overallState.equalsIgnoreCase(UNKNOWN)) {
-                                overallState = CLOSED;
-                            }
-                        }
-                    }
-                }
-            } catch (IllegalArgumentException | IllegalAccessException e) {
-                return UNKNOWN;
-            }
-        }
-        return overallState;
-    }
 }
